pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.42",
  "requests>=2.32",
  "markdown>=3.10",
  "boto3-stubs[s3]>=1.42.62",
  "ffmpeg-python>=0.2.0",
  "pydantic>=2.12.5",
  "pygithub>=2.8.1",
  "python-dotenvx>=0.3.0",
]

[tool.uv]
dev-dependencies = []

[tool.pyright]
venvPath = "."
venv = ".venv"
